Remove CpsAssetPageMain, move logic to /pages
